requirements.txt
pytest
playwright
pytest-playwright
appium-python-client
python-dotenv
pytest-html